agents-api/requirements.txt
openai-agents
fastapi
uvicorn
sqlalchemy
psycopg2
pymilvus
pydantic
pydantic-settings
rank_bm25
pandas
openai
python-dotenv
tqdm
presidio-analyzer
presidio-anonymizer
llmlingua
langsmith
deepeval
asyncpg
httpx
